Refactor + Update Tests + Update Docs
