+++ tests/test_dfs/dfs_cleanup.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+dmg -i pool destroy -f TEST_POOL
