--- tests/test_dfs/dfs_test.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-dmg -i pool create -s 2G TEST_POOL
-daos cont create --pool=TEST_POOL --type=POSIX TEST_CONT
